Refactor docs to one sentence per line
This is a borrowed convention from AsciiDoc that I fell in love. Writing docs
as one sentence per line is weird as a writer, but it makes a PR against docs
easier to read and understand. The diff will align sentences side-by-side. So
this commit brings our existing docs into this format.

Some of the docs up to this point were already written this way, so this also
makes everything consistent too.