Add new document to explain GitHub workflow (how to add and maintain projects)
This commit begins a new document to explain and document our workflow on
GitHub. This document is intended for project maintainers at UNICEF. It is a
list of things to do whenever making a new GitHub repo, maintaining one, and
how to communicate about our projects. It is intended as a cheat-sheet or
reference for later.

Feel free to contact @jflory7 about the content here at any time, even if it's
after the end of the internship.