:memo: README: Update current status, fix broken links, formatting
This commit is a minor refresh of the MagicBox repository README, which
is one of the most visible GitHub repositories by UNICEF. This adds a
note to the top that says that this repository is no longer actively
maintained. It also fixes some broken links, and moves the README to
follow the one-line-per-sentence convention for technical writing.

I was compelled to add a note because although I am not up-to-date on
the latest news with MagicBox, Project Connect, or Giga, I felt
obligated to make an update as a past committer and manager of this
repository!

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>